+++ lib/rubocop/cop/style/class_vars.rb
@@ -6,7 +6,7 @@ module Rubocop
       # This cop checks for uses of class variables. Offences
       # are signaled only on assignment to class variables to
       # reduced the number of offences that would be reported.
-      class AvoidClassVars < Cop
+      class ClassVars < Cop
         MSG = 'Replace class var %s with a class instance var.'
 
         def on_cvasgn(node)